Renamed `LibraryFileTree.java` to `LibraryFile.java` and `LibraryFileNode.java` to `ReactorNode.java` to better reflect their functionality; Improved documentation across the renamed files for clarity.
