+++ lsp/src/main/java/org/lflang/diagram/lsp/LFLanguageServerExtension.java
@@ -105,10 +105,12 @@ class LFLanguageServerExtension implements ILanguageServerExtension {
    * {@code filePath}.
    *
    * @param filePath the URI of the LF file of interest
-   * @return A message describing the outcome of the build process.
+   * @return A {@code CompletableFuture<LibraryFile>} representing the asynchronous computation * of
+   *     the parsed reactor structure. If an error occurs during parsing, the future will * complete
+   *     with {@code null}.
    */
   @JsonRequest("generator/getLibraryReactors")
-  public CompletableFuture<LibraryFileTree> getLibraryReactors(String filePath) {
+  public CompletableFuture<LibraryFile> getLibraryReactors(String filePath) {
     return CompletableFuture.supplyAsync(
         () -> {
           try {
@@ -126,14 +128,15 @@ class LFLanguageServerExtension implements ILanguageServerExtension {
    * Retrieve the target position specified in the LF program file at the given path.
    *
    * @param path The path to the LF program file.
-   * @return A CompletableFuture containing the NodePosition object representing the position of the
-   *     target, or null if an error occurs during parsing or if the target position is not found.
+   * @return A {@code CompletableFuture<NodePosition>} containing the NodePosition object
+   *     representing the position of the target, or null if an error occurs during parsing or if
+   *     the target position is not found.
    */
   @JsonRequest("generator/getTargetPosition")
   public CompletableFuture<NodePosition> getTargetPosition(String path) {
     return CompletableFuture.supplyAsync(
         () -> {
-          NodePosition targetPosition = null;
+          NodePosition targetPosition;
           try {
             URI uri = URI.createURI(path);
             // LF program file parsing
@@ -151,14 +154,14 @@ class LFLanguageServerExtension implements ILanguageServerExtension {
 
   /**
    * Parse a library of reactors specified by the provided URI and construct a hierarchical
-   * libraryFileTree representation.
+   * libraryFile representation.
    *
    * @param uri The URI specifying the location of the library.
-   * @return A Tree object representing the hierarchical structure of the reactor library, or null
-   *     if an error occurs during parsing.
+   * @return A {@code LibraryFile} object representing the hierarchical structure of the reactor
+   *     library, or {@code null} if an error occurs during parsing.
    */
-  public LibraryFileTree parseLibraryReactors(URI uri) {
-    LibraryFileTree res = new LibraryFileTree(uri.toString());
+  public LibraryFile parseLibraryReactors(URI uri) {
+    LibraryFile res = new LibraryFile(uri.toString());
     try {
       Resource resource = getXtextResourceSet(uri).getResource(uri, true);
       Model m = (Model) resource.getContents().get(0);
@@ -168,7 +171,7 @@ class LFLanguageServerExtension implements ILanguageServerExtension {
           r -> {
             INode node = NodeModelUtils.getNode(r);
             NodePosition nodePosition = new NodePosition(node.getStartLine(), node.getEndLine());
-            res.addChild(new LibraryFileTreeNode(r.getName(), res.getUri(), nodePosition));
+            res.getChildren().add(new ReactorNode(r.getName(), res.getUri(), nodePosition));
           });
     } catch (Exception e) {
       return null;
